Move AddLink to incubating module
