+++ sdk/trace/src/main/java/io/opentelemetry/sdk/trace/SdkSpan.java
@@ -13,6 +13,7 @@ import io.opentelemetry.api.trace.SpanContext;
 import io.opentelemetry.api.trace.SpanKind;
 import io.opentelemetry.api.trace.StatusCode;
 import io.opentelemetry.context.Context;
+import io.opentelemetry.extension.incubator.trace.ExtendedSpan;
 import io.opentelemetry.sdk.common.Clock;
 import io.opentelemetry.sdk.common.InstrumentationScopeInfo;
 import io.opentelemetry.sdk.internal.AttributeUtil;
@@ -35,7 +36,7 @@ import javax.annotation.concurrent.ThreadSafe;
 
 /** Implementation for the {@link Span} class that records trace events. */
 @ThreadSafe
-final class SdkSpan implements ReadWriteSpan {
+final class SdkSpan implements ReadWriteSpan, ExtendedSpan {
 
   private static final Logger logger = Logger.getLogger(SdkSpan.class.getName());
